Hello World!
This is the first commit for Upshift (but it is not its first step).
May peace prevail on Earth. May human beings collaborate with one
another. May we work hand-in-hand to solve the problems of our time.
May we never stop dreaming of a better future and never cease working
towards its realization. May we never lose hope in the possibility of
Utopia. May we never take 'impossible' for an answer and create
possibility even when it seems out of option. May we forgive ourselves
and others. May we celebrate this magic that are our lives on this
planet.

With peace. With love. And with happiness.