Add model handle
Handles are used for users and teams to act as a unique identifier. They
are more commonly known as usernames but here we use the term 'handle'
to make clear that users as well as other models (teams, in the future)
share the same database table for their identifiers (handles).