Use route key 'profiles' for users
Enable routing a user object to profile with the url_for helper. url_for
bases its generated path on the route key of the object that it is
passed. This failed for user objects because we do not have a user_path
route. Manually setting the route key for users to 'profiles' allows
user objects to use the profile_path route.