Fix: VersionControl::File#content is nil for new instances
When calling VersionControl::File#content on instances where @content is
nil and @oid is nil, an error would occur. This was due to the fact that
Rugged::Repository#lookup raises an error if a non-string parameter is
passed. This fixes the problem by only calling lookup if
VersionControl::File#oid is not nil.