Validate: When updating file, exactly one attribute must change
Before a new VersionControl::File revision is saved, it is validated
that exactly one attribute (name or content) has changed. If both have
changed, it adds an error to alert the user that only one attribute may
change per revision. If neither have changed, it adds an error to alert
the user that at least one change must be made to the file.