Discussion: Redirect on type-ID mismatch
When a user requests a discussion, such as ../suggestions/4, but the
type (suggestions) does not match the type of the discussion with that
scoped id (4), then redirect the user to the path with the corrected
type.
Additionally, a user can request a discussion by scoped it without
knowing its type by simply requesting ../discussions/:scoped_id.