Merge class Handle into Profiles::Base
Handle class was used to manage identifiers (usernames and team
usernames). This was necessary because we had two separate classes (User
and Team) that were both sharing handles. So we used a single table for
handles. Since both users and teams are now using the Profile table,
we can just directly store the handle on that table.