Add support for references in DelayedJobs
When creating an ActiveJob, you can pass a hash including a reference
to a model instance that will then be saved to the job. For example:
TestJob.perform_later(reference: user) will result in the columns
delayed_reference_type and delayed_reference_id being stored alongside
the job in the database.
This makes it easy to track for 'whom' a job is being executed.