Revert "Create notification channel on file#create"
Revert the notification channel commit because there is a limit for
channel subscriptions -- and it seems rather low, maybe 100/hour. The
actual rate limit and the time frame are unclear because Google does
not publish this quota. In any case, we would need to do 1000+
subscriptions/hour if we were to subscribe to every file.
For more info read here:
https://stackoverflow.com/a/38326165/

So instead we're looking at a new approach (manually pinging for
changes) that does not make use of notification channels.

This reverts commit 194a76ea018d45f8283ddee72904ab4a10e10e2f.