Use object ID for repository path (to avoid renames)
Instead of using a repository path that is subject to change (such as
/user-handle/project-slug), we instead store repositories at a permanent
location consisting of the object's ID. That way, we never have to
move repository files around due to renaming. The fewer read/write
operations we have to perform on repositories, the smaller the
possibility of race conditions.