Suppress log output for rendering of views
Since we are heavily relying on partials in our views and each partial
ordinarily generates one line of output in the log, our log files is
starting to be overloaded with rather useless information about how
long it took to render each individual partial.

We can suppress this information and our log will still list the total
time it took to render the view including all partials.