Fix: Display all file diff changes
When displaying diff/change for files, show indicator for each change.
This is currently relevant only for files that have been both modified
and moved because these are the only two changes that can occur
together. Other changes are mutually exclusive.
That said, in the future, there may be more possible combinations (such
as renamed).