Fix: Ignore changes to mime type when updating files
Changing the mime type has serious implications if it means switching
from a folder to a file or vice versa. For example, what should happen
when a folder with lots of child files suddenly takes on a non-folder
mime type? To avoid it all, once assigned, we keep mime types constant.