Fix: RevisionDiff#changed_files_as_diff
When generating deltas by diffing two trees, the returned deltas include
deltas not just for files that changed blob oid or parent id. All files
that changed path are included in the details. For example, if a file's
parent moves but the file itself has the same parent_id and blob, it
will show up in the list of details because its relative path has
changed. To avoid this, for now, we just run select(&:changed?) on the
array of file diffs at the end.