Refresh Drive authorization token before every spec
API connections need to refresh their authorization token every hour.
If the token is expired, the refreshing is done automatically by the
Google API gem. But this leads to inconsistent VCR recordings where
sometimes the token is refreshed at the start of the spec, and other
times it is not. To get consistent recordings, we force refresh the
token at the beginning of every spec using the GoogleDriveTestHelper.