Drive API: Add #find_file(id) and #find_file!(id)
Replace #fetch_file(id) method with #find_file!(id) method. This is done
to mirror the design of ActiveRecord finder methods. #find_file! will
raise an error if the file is not found. The non-bang version #find_file
can be used to return the file, or nil if not found.