Simplify application infrastructure
Move from single to multi source model, where each repository contains
its own set of file records (as opposed to global file resources shared
between repositories). The reason for the multi source decision is that
this makes it possible to have different versions of the same file in
different repositories (which may be necessary due to two files having
different owners who have different levels of file access).

We decided not to touch existing models, so that we can get the new
infrastructure working, then deploy, and then manually migrate data from
the single to the multi source model. That is easiest if the existing
models are not touched.

We are also remodeling the infrastructure to prepare for multi branch
support and backups.