Fix: ForceSyncsController: Pull children of freshly deleted folders
When a folder was removed in the process of a force sync, it would no
longer return true to #folder? (because its mime type changed to nil).
This is fixed by querying #folder_now_or_before_last_save? which returns
true as long as the mime type is or was folder.