Rubocop: Do not enforce FrozenStringLiteralComment in migrations
Migrations tend not to have strings and, if they do, these strings tend
not to be modified. Therefore, we can safely turn off the enforcement
of the frozen string literal comment. This saves us some effort and
means we do not have to touch every migration when Ruby v3.0 launches.