Add model VCS::Content for storing file content
An instance of VCS::Content represents one content version of a file.
We want to move away from having the remote file's content version ID
be the determinant of whether the file was modified because when we
replace a file (as part of restoration or merge), the file's external
ID changes and its remote content version resets to 1. Two file
snapshots could have completely different external IDs and version IDs,
but represent the same content.