Add model VCS::RemoteContent for mapping remotes to content
An instance of VCS::RemoteContent maps a remote file ID and its remote
content version ID to a local instance of VCS::Content. This allows us
to map multiple remote files to the same content version, as would be
the case in a file restore or in a merge (because both of them work by
replacing the remote file with a new file).