Capistrano: Fix 500 page generation
When generating the 500 error page, curl is called on the hostname with
path 500. In our case, the hostname is openly.one but the error page
lives at www.openly.one. Curl did not follow the redirect, so the 500
page became a 'moved permanently' message. Add flag -L to instruct curl
to follow redirects.

[ci skip] because no testable files were touched.