Add Contribution model
The contribution model represents suggestions of changes to projects.
This is first step towards giving users a streamlined process for
proposing edits in a project. When a user creates a new contribution,
they receive a copy of each file in the project that they can edit. Any
edits to those files will show up in the contribution only. Maintainers
of the project can then decide if they want to accept the contribution.
If they do, all changes will be copied over to the project.