Rename VCS::StagedFile to VCS::FileInBranch
A VCS::FileInBranch record is a relationship between a VCS::File and a
VCS::Branch. Remove all mentions of 'stage' or 'staged'. The concept
of stage does not really apply in our case, as in Git it is meant to
refer to files that have been staged for commit (as opposed to files in
the working directory). So we are just calling everything a file. To
differentiate repository-wide files from the branch-specific files,
we are calling the branch-specific files FileInBranch.