Rename VCS::FileRecord to VCS::File
We originally used the model name FileRecord when it was used in the
toplevel namespace (and thus File would have conflicted with Ruby's
default File class). Since moving to the VCS namespace, we no longer
need to worry about conflicts with the File class and can rename
VCS::FileRecord to VCS::File.