VCS::FileInBranch: Add finder for hashed file ID or remote file ID
While we are migrating to hashed file IDs as the primary IDs for
navigation, we still want to support navigation by remote file ID. Why?
To allow someone to find their Google Drive folder on Openly by typing
/handle/slug/folders/google-drive-id or
/handle/slug/files/google-drive-id/info. But we will use hashed file IDs
as the primary navigation because they are persistent (continue to work
even when a file is replaced or restored) and function across revisions
and branches.