VCS::ContentDiffer: Improve diffing algorithm
We want to track changes to spaces and new lines. In order to do so,
we cannot use either of the two as dwdiff --white-space delimiters.
Instead, we want to use a special character that we do not care about
and that will not appear in user-generated content: \b (backspace)
We wrap all whitespace in this custom whitespace delimiter to treat
whitespace as an individual word.

The downside of this is that dwdiff will frequently NOT group related
changes together, but rather have each word represent its own change
(with the white space in between being retained). To avoid having lots
of changes that would be hard to read, we aggregate all changes that
are separated by a whitespace.