File Content Changes: Use 'After' header for captured changes
When viewing a captured content change side-by-side, do not use 'Now' as
the header label - it can be confusing because it sounds like 'now' as
in the current moment. Use 'After' instead because this makes it clearer
that the text shown represents the document at/after the point of the
revision.