Fix force-sync folder: Pulls all children not visible in folder view
When force-syncing a folder, the #pull_children method will not call
the #pull method on any children that are already present in branch and
can be manually force-synced by the user. It does #pull new children
as well as children without a version because those are not visible in
the folder view and cannot be manually force synced by the user.

Fixes [#247](https://github.com/OpenlyOne/openly/issues/247).