Add feature spec for visitors
Add feature specs to test correct functionality of projects for
visitors. Visitors are either guests (users who are not signed in) or
non-collaborators (users who are signed in but not collaborators of a
project).

The features tested are visibility of public projects, read access to
the projects' archives, removal of read access to the archive upon
making a project private, and showing of the project's last revision
(rather than the staged files) when clicking on 'Files'.