Travis: Add explicit instructions to install bundler
Our Travis builds were failing with the message:
```
The program 'bundle' is currently not installed. To run 'bundle' please
ask your administrator to install the package 'bundler'
```

It seems that this may have to do with the release of Bundler 2.0:
https://bundler.io/blog/2019/01/03/announcing-bundler-2.html)
but I'm not sure.

In any case, I saw that the .travis.yml in the Rails project runs
`gem install bundler`, so I copied that.