Admin Project Dashboard: Do not search tag_list
When searching the project dashboard, Rails has been throwing an
internal error. This is due to the fact that Administrate is attempting
to search the attribute `tag_list` in the database, but `tag_list` is an
entirely virtual attribute (built from the tags array). To avoid the
error, `tag_list` is explicitly removed from the fields to search.