VCS: Generate content_id in FileInBranch rather than Version
To generate a content_id, the attributes `repository`, `remote_file_id`,
and `content_version` are required. These attributes are naturally
present on the FileInBranch rather than the Version. Therefore, it
makes sense to generate the content_id on the FileInBranch model and
then pass the generated content_id along to Version.