VCS::Version: Remove content_version column
The VCS::Version should not reference a specific remote content version
as two different remote files with different remote content versions
could be referring to the same VCS::Version instance (as is the case
when a file restore replacement or addition is performed).

The same is true for the remote_file_id attribute of VCS::Version which
we will remove in the next commit.