VCS::Version: Remove remote_file_id column
Versions should not track the remote file ID. Instead, FileInBranch
should carry the sole responsibility for doing so. That is because one
and the same version may belong to many different FileInBranches that
all have different remote file IDs. It would be strange to have the
version itself associated with one of those remote file IDs
specifically.