VCS::Branch: When forking, copy all committed files from origin
When creating a fork of a branch named `origin`, copy all committed
files from `origin` to the new fork. This ensures that all files are
present in the fork: Even if the remote file restoration process were
to fail, the files would correctly be marked as deleted in the fork and
the user could manually restore files one by one.