Contribution: Users can review suggested changes
On a contribution's review tab in a contribution, users can see a list
of suggested changes. The suggested changes are the changes between the
last commit on the project's master branch and the contribution's files.
This is meant to help a project maintainer evaluate the consequences of
accepting the contribution.