Fix: Feature specs
Refactor the `create_revision` methods defined in a few of the feature
specs we use. It looks like these methods are not locally scoped to
just the feature they are defined in but shared across other feature
specs as well. This can lead to issue when the same method accept
different arguments as was the case with `create_revision` where some
methods require a title and some did not. I've changed all
`create_revision` methods to require a title.

It should be noted that this is a temporary fix in any case. Long term,
we should move to not having methods defined in global scope and instead
use helpers (or rich factories that handle the revision create and
publication).