Update project preview cards
Use the new, minimalist style for the project preview cards. This
focuses exclusively on title and description and removes background
image and tags.

That is because background images are not used (users
cannot edit those). Tags could be used but are not very much used and
only a few tags would fit on the card in any case and the value-add of
the tags is questionable.

Resolves [#291](https://github.com/OpenlyOne/openly/issues/291).