Fix VCS::FileInBranch factory: Make :unchanged work with :create
Improve the VCS::FileInBranch factory, so that the :unchanged trait
also works when using the FactoryBot :create strategy. Until now, the
persistence of the record would overwrite the current_version assigned
by the factory. This commit fixes that by writing the file's actual
attributes to the assigned version, so that the version does not change
on save.