Profile page: Show only projects that user can view
On the profile page, show only projects that the user can view (such as
public projects or private projects where the user is a collaborator).
Private projects that the user cannot view are no longer shown. That is
good because they were meant to be private in the first place.

Resolves [#302](https://github.com/OpenlyOne/openly/issues/302).