Contribution: Calculate suggested changes relative to origin revision
When preparing the revision for accepting a contribution, calculate
the revision by first calculating the file diffs between contribution
and origin revision. Then replay those file diffs on top of the latest
revision in the project. This leaves any changes made on master since
starting the contribution untouched, unless the same files are
overwritten in the contribution.