Fix Contribution: Mark new files as committed after acceptance
When adding a new file in a contribution and then accepting the
contribution, the new file would be copied over to the master branch,
but not marked as committed. Instead, the file would incorrectly show
up as added.

The reason is that the file's `committed_version_id` is nil. When we run
the `VCS::Commit#update_files_in_branch` to update the
`committed_version_id` of all files in the branch, the new file is not
yet present in the branch. That's because the FileRestore job runs in
the background and is only executed **after** the commit is published
and `#update_files_in_branch` has been run.

To fix this, we copy over any untracked files from the commit to the
branch **before** running `#update_files_in_branch`. All of this logic
is contained in a new `VCS::Branch.mark_files_as_committed(commit)`
method that first runs `#copy_untracked_files_from_commit` and then
runs `#copy_committed_version_id_from_commit`.

Fixes [#328](https://github.com/OpenlyOne/openly/issues/328).