Contribution replies: Return to reply form if errors during create
Specify anchor 'new-reply' on the form for a new contribution reply.
This returns the user to the form (bottom of the discussion thread) if
an error is encountered during the reply creation. It ensures that the
user sees that an error was encountered which would not otherwise be
obvious, especially when there is a long list of replies.