Henkei: Avoid false linebreaks in plain text extraction
The previous method of plain text extraction from documents was causing
incorrect line breaks to be inserted into text.

The problem is that we're currently extracting plain text by first using
Apache Tika's HTML extraction and then using Nokogiri's HTML to text
method. That's because Tika's text extraction includes unwanted
artifacts, such as `[image: ...]` and `[bookmark: ...]`.
See [the TIKA-2755](https://jira.apache.org/jira/browse/TIKA-2755).

The solution to this is to use the Tika server's `/rmeta/text` endpoint
(as opposed to `/tika`) and then parse the JSON response. The plain text
content is in the "X-TIKA:content" field.

This field includes a bunch of line breaks at the beginning of the
document. These need to be stripped. In a future improvement, it would
be great to strip only those linebreaks that are not part of the
document.

Resolves [#342](https://github.com/OpenlyOne/openly/issues/342)